// src/interfaces/auth/auth.interfaces.ts
export interface SignInResponse {
	user: UserLoginData;
	token: string;
}

export interface UserLoginData {
	id: string;
	email: string;
	fullName: string;
	status: string;
	roles: string[];
	insurance?: InsuranceRes | null;
}

// Generated by https://quicktype.io

export interface InsuranceRes {
	id: string;
	description: string;
	damage_cover_liability: number;
	hourly_rate: number;
	daily_cap: number;
	status: string;
	created_at: string;
	updated_at: string;
}

// Generated by https://quicktype.io

export interface ValidateTokenResponse extends SignInResponse {
	id: string;
	email: string;
	fullName: string;
	roles?: string[];
	statusCode?: number;
	message?: string;
	error?: string;
}